LayoutTests/platform/qt-5.0/http/tests/security/xss-DENIED-xsl-external-entity-redirect-expected.txt

CONSOLE MESSAGE: Unsafe attempt to load URL  from frame with URL http://127.0.0.1:8000/security/xss-DENIED-xsl-external-entity-redirect.xml. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe attempt to load URL  from frame with URL http://127.0.0.1:8000/security/xss-DENIED-xsl-external-entity-redirect.xml. Domains, protocols and ports must match.

This test includes a cross-origin external entity. It passes if the load fails and thus there is no text below this line.
